shader_runner/spirv: skip for GL extensions without SPIR-V equivalent
There are some GL extensions/features, like shader subroutines, that
doesn't have their equivalent on SPIR-V, so those tests should be
skipped on SPIR-V mode. We reuse the code that checks if a given
SPIR-V extension equivalent to a GL extension is supported, by setting
the mapping to a string that will never be the name of a SPIR-V
extension.

Acked-by: Marek Olšák <marek.olsak@amd.com>
Acked-by: Timothy Arceri <tarceri@itsqueeze.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/1>