replay/apitrace: Forward logs when the test crashes
Currently, we only dump the traceback and the return code outputted from
the inner command. We should provide the logs (stdout + stderr) in this
case as well, to make it easier to investigate what is happening,
without the need to reproduce the inner command manually.

Parametrize test_from_yaml_\w*_traces? tests. Only check the last output
line in this test, as we now are little more verbose with the output.

Signed-off-by: Guilherme Gallo <guilherme.gallo@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/777>