gles3: fix TF to UBO layout
This test reads a transform feedback buffer as an UBO, but it doesn't
specify the layout.

With some drivers it is applied with std430 layout (and thus it works),
but with other it applies std140 layout. The problem with this layout is
that each element of the array must be aligned with the size of vec4.
Hence, reading it as an array of integers will fail.

To provide coherence among all drivers, let's declare the UBO with
std140, and read it as ivec4 instead of integer, so we can read each
value.

Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>
Reviewed-by: Emma Anholt <emma@anholt.net>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/800>