nv_copy_image-simple: Use red instead of green
This demonstrates an issue with the iris driver. It does not account for
the automatic fast clearing behavior seen on TGL when copy operations
are performed.

Copying red onto dark red is enough to engage the automatic clearing
behavior because hardware is reinterpreting the floating point clear
values to the format chosen by BLORP (UINT), then seemingly clamping to
the 8bit-per-channel maximum of 255. So, the source value of 1.0 becomes
255 and the destination clear value of 0.5 becomes 255. Because the
pixels written match the destination clear value and the area is large
enough, some (if not all) get converted to fast cleared blocks.

Ref: https://gitlab.freedesktop.org/mesa/mesa/-/issues/3732
Acked-by: Emma Anholt <emma@anholt.net>
Acked-by: Rohan Garg <rohan.garg@intel.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/808>