modifiers: Remove sampler-view tests
When a particular modifier is forced, it is not always possible to do
various format casts.  Normally the gl driver would internally demote
the modifier, but this isn't possible with imported buffers with a
particular modifier forced.

Perhaps this is a spec issue, but since it is impossible in some cases,
so lets just remove the tests.  The alternative would be to somehow
maintain an allow list of which format casts are permitted for a given
modifier, which doesn't seem reasonable.

Signed-off-by: Rob Clark <robdclark@chromium.org>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/849>