properly check for libgen.h
Some users are reporting that basename is not available when building on
Linux with musl libc. And since the POSIX spec[1] says that basename is
defined in libgen.h, we should include that when available.

So let's properly detect the header, and include it if it exists. This
should hopefully make things a bit more robust.

Since we're also including this from the CL program-tester, let's
rearrange the includes a bit so we know that config.h has been included.

Reviewed-by: David Heidelberg <david.heidelberg@collabora.com>
Tested-by: David Heidelberg <david.heidelberg@collabora.com>
Reviewed-by: Khem Raj <raj.khem@gmail.com>
Tested-by: Khem Raj <raj.khem@gmail.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/888>