masked-clear: test depth/stencil clear cross contamination
Add a test to validate that glClear with only depth or stencil buffer
bits set only affect the respective portion of the ZS buffer. This
reproduces a bug in etnaviv, where such clears would affect the whole
buffer when fast-clear is in use.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/902>