glx: don't expect a glx error
The spec says BadMatch should be generated, which is a X error not a GLX one.

The original code was only successful if "glx_error_code != -1" (= there
was a GLX error) and "validate_glx_error_code(BadMatch, -1)" (= there was
a X error and no GLX error) which is wrong.

This commit the glx_error_code != -1 test, but still verifies that BadMatch
is generated.

This fixes this test on radeonsi (and probably other drivers).

Reviewed-by: Marek Olšák <marek.olsak@amd.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/981>