glx: don't fail test if the X server is inconsistent
This bug is going to be fixed in server side, but failing
the piglit test in this case doesn't make much sense as the
driver cannot do anything about the failure.

In this situation, the test would now report a warning to
be able to tell the difference between a client-side failure
and a server-side one.

Reviewed-by: Marek Olšák <marek.olsak@amd.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/983>