OVR_multiview: Add fb_incomplete_view_range test
Add a test for the additional conditions required for framebuffer
multiview attachment completeness in the OVR_multiview spec, namely that
all the selected layers are less than the layer count of the texture.

We test with a 4 layer texture, and an attachment of layers 3-4 (which
should be considered an incomplete attachment), and attachments of
layers 0-2 and 2-3 (which should be considered complete).

Signed-off-by: James Hogan <james@albanarts.com>
Acked-by: Marek Olšák <marek.olsak@amd.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/987>