OVR_multiview: Add fb_blit_first_view test
Add test that blits to multiview framebuffers only blit to the first
view in the framebuffer.

A normal 2D read framebuffer containing green is blitted to a multiview
draw framebuffer containing red, either as a single full blit or two
half blits, and the multiview framebuffer is read to confirm that only
the first view is changed to green and the others remain red.

This test is called with different arguments to control whether full
blits or partial blits are performed, and the base view layer index.

Signed-off-by: James Hogan <james@albanarts.com>
Acked-by: Marek Olšák <marek.olsak@amd.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/987>