arb_direct_state_access: fix internaformat query behavior
The error condition and error value changes with and without
GL_ARB_internalformat_query. Even though this test are using the named
version of these APIs, they inherit their error-handling rules from the
non-named versions.

Before introducing that extension, we need to check against
GL_MAX_SAMPLES, which is what we did. But the reported error should be
GL_INVALID_VALUE, not GL_INVALID_OPERATION like we checked against.

To quote the OpenGL 4.1 spec, section 4.4.2 ("Attaching Images to
Framebuffer Objects"):

   If either width or height is greater than the value of
   MAX_RENDERBUFFER_SIZE, or if samples is greater than the value of
   MAX_SAMPLES, then the error INVALID_VALUE is generated.

After introducing that extension, we need to check against the
per-internal-format max instead of the "global max", because this is
now the smallest value that is *guaranteed* to work instead.

To quote the OpenGL 4.1 spec, section 4.4.2 ("Attaching Images to
Framebuffer Objects"):

   If either width or height is greater than the value of
   MAX_RENDERBUFFER_SIZE, then the error INVALID_VALUE is generated.
   If samples is greater than the maximum number of samples supported
   for internalformat, then the error INVALID_OPERATION is generated
   (see GetInternalformativ in section 6.1.15).

So, both the value to check against and the error code changed at the
same time. The GL_ARB_internalformat_query specification also contains
these changes. It seems we never tested something that was consistent
here, but was instead relying on the max samples for GL_RGBA being the
same as the formatless max samples, which is not guaranteed to be the
case.

Reviewed-by: Marek Olšák <marek.olsak@amd.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/1003>