tests/util: provide simple array for cleanup handlers
This makes it possible to call piglit_set_destroy_func multiple times.
This is done so that tests can register cleanup handler as well and
avoid usage of atexit which can cause issues with frameworks such as the
vulkan validation layers.

Signed-off-by: Tapani Pälli <tapani.palli@intel.com>
Acked-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/1011>