glslparser: added test for accidental token pasting
If after macro expansion two identifiers would appear one
after the other, there should be a space between them so
that they aren't mistaken for one identifier. That is,
token pasting should only happen with a deliberate `##`,
not by accident. Both clang and gcc insert a space in the
test here, but Mesa did not (pending a merge request to
fix this).

Reviewed-by: Erik Faye-Lund <erik.faye-lund@collabora.com>
Reviewed-by: Marek Olšák <marek.olsak@amd.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/1033>