glsl-1.40: require GL 4.5 for the two-stages test
This test is requiring GL >= 3.1, but then it declares to use GL 4.5, so
that makes the test failing for drivers supporting GL 3.1 but not 4.5.

We rely on using interface blocks and bindings, so we can't just declare
a 3.1 version.

So instead, let's require GL 4.5 instead, so those drivers just skip
this instead of failing.

Fixes: 2cc580621 (glsl-1.40: add a UBO test that I don't remember why I needed it)
Reviewed-by: Marek Olšák <marek.olsak@amd.com>
Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/1034>