pixel local storage: make some format tests use unnamed structures
Pixel local storage variables can be declared in two different ways:

__pixel_localEXT mydata1 {
   vec4 color1;
   vec4 val1;
} pls1;

__pixel_localEXT mydata2 {
   vec4 color2;
   vec4 val2;
};

In the first way the variables are referred to in later code as
pls1.color1 and pls1.val1, but in the second way they are just color2
and val2. The final generated code at the device end should be the same
for both cases, but internally the compiler handles them quite differently
(the first is a single variable that is a struct with fields, the second
has two separate variables that are vec4). This commit makes sure that
both paths through the compiler are tested.

Reviewed-by: Erik Faye-Lund <erik.faye-lund@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/1054>